@@ -7,7 +7,7 @@
     <groupId>org.monarchinitiative.threes</groupId>
     <artifactId>3S</artifactId>
     <packaging>pom</packaging>
-    <version>1.0.4</version>
+    <version>1.1.0</version>
     <modules>
         <module>threes-core</module>
         <module>threes-ingest</module>
@@ -19,10 +19,11 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <htsjdk.version>2.19.0</htsjdk.version>
-        <jannovar.version>0.28</jannovar.version>
-        <slf4j.version>1.7.25</slf4j.version>
+        <jannovar.version>0.29</jannovar.version>
+        <slf4j.version>1.7.29</slf4j.version>
+        <hyperutil.version>0.1.2</hyperutil.version>
         <protobuf.version>3.5.1</protobuf.version>
         <guava.version>27.1-jre</guava.version>
         <junit-jupiter.version>5.4.1</junit-jupiter.version>
@@ -50,6 +51,12 @@
     </parent>
 
     <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-configuration-processor</artifactId>
+            <optional>true</optional>
+        </dependency>
+
         <!--        Libraries for testing         -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
