@@ -1,43 +1,41 @@
 package org.monarchinitiative.threes.core.model;
 
+import de.charite.compbio.jannovar.reference.GenomeInterval;
+
+import java.util.Objects;
+
 /**
  *
  */
 public abstract class SplicingRegion {
 
-    protected final int begin;
+    protected final GenomeInterval interval;
 
-    protected final int end;
-
-    protected SplicingRegion(int begin, int end) {
-        this.begin = begin;
-        this.end = end;
+    protected SplicingRegion(GenomeInterval interval) {
+        this.interval = interval;
     }
 
-    public int getBegin() {
-        return begin;
-    }
-
-    public int getEnd() {
-        return end;
+    public GenomeInterval getInterval() {
+        return interval;
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof SplicingRegion)) return false;
-
+        if (o == null || getClass() != o.getClass()) return false;
         SplicingRegion that = (SplicingRegion) o;
-
-        if (begin != that.begin) return false;
-        return end == that.end;
-
+        return Objects.equals(interval, that.interval);
     }
 
     @Override
     public int hashCode() {
-        int result = begin;
-        result = 31 * result + end;
-        return result;
+        return Objects.hash(interval);
+    }
+
+    @Override
+    public String toString() {
+        return "SplicingRegion{" +
+                "interval=" + interval +
+                '}';
     }
 }
