@@ -1,24 +1,25 @@
 package org.monarchinitiative.threes.core.model;
 
-import org.monarchinitiative.threes.core.reference.fasta.InvalidCoordinatesException;
+import de.charite.compbio.jannovar.reference.GenomeVariant;
+import xyz.ielis.hyperutil.reference.fasta.SequenceInterval;
 
 /**
- * POJO for {@link SplicingVariant}, {@link SplicingRegion} (either exon or intron), and {@link SequenceInterval}, all
+ * POJO for {@link GenomeVariant}, {@link SplicingRegion} (either exon or intron), and {@link SequenceInterval}, all
  * adjusted to strand of the transcript.
  */
 public class SplicingTernate {
 
-    private final SplicingVariant variant;
+    private final GenomeVariant variant;
 
     private final SplicingRegion region;
 
     private final SequenceInterval sequenceInterval;
 
-    private SplicingTernate(SplicingVariant variant, SplicingRegion region, SequenceInterval sequenceInterval) {
-        if (variant.getCoordinates().getStrand() != sequenceInterval.getCoordinates().getStrand()) {
-            throw new InvalidCoordinatesException(String.format("Sequence %s and variant %s are not on the same strand", sequenceInterval, variant));
+    private SplicingTernate(GenomeVariant variant, SplicingRegion region, SequenceInterval sequenceInterval) {
+        if (variant.getChr() != region.getInterval().getChr() || variant.getChr() != sequenceInterval.getInterval().getChr()) {
+            throw new IllegalArgumentException(String.format("Creating ternate from data on different contigs - variant: `%d`, region: `%d`, sequence: `%d`",
+                    variant.getChr(), region.getInterval().getChr(), sequenceInterval.getInterval().getChr()));
         }
-
         this.variant = variant;
         this.region = region;
         this.sequenceInterval = sequenceInterval;
@@ -28,16 +29,16 @@ public class SplicingTernate {
     /**
      * Make the ternate from variant, region, and sequence interval - all of them must be on the same chromosome strand.
      *
-     * @param variant          {@link SplicingVariant} variant
+     * @param variant          {@link GenomeVariant} variant
      * @param region           {@link SplicingRegion} region
      * @param sequenceInterval {@link SequenceInterval} sequence interval
      * @return ternate
      */
-    public static SplicingTernate of(SplicingVariant variant, SplicingRegion region, SequenceInterval sequenceInterval) {
+    public static SplicingTernate of(GenomeVariant variant, SplicingRegion region, SequenceInterval sequenceInterval) {
         return new SplicingTernate(variant, region, sequenceInterval);
     }
 
-    public SplicingVariant getVariant() {
+    public GenomeVariant getVariant() {
         return variant;
     }
 
