@@ -1,5 +1,9 @@
 package org.monarchinitiative.threes.core.reference;
 
+import de.charite.compbio.jannovar.reference.GenomePosition;
+
+import java.util.Optional;
+
 /**
  * POJO for grouping location data of variant with respect to transcript.
  */
@@ -7,26 +11,37 @@ public class SplicingLocationData {
 
     private final SplicingPosition position;
 
+    private final GenomePosition donorBoundary;
+    private final GenomePosition acceptorBoundary;
     private final int intronIdx;
-
     private final int exonIdx;
 
     private SplicingLocationData(Builder builder) {
         intronIdx = builder.intronIdx;
         exonIdx = builder.exonIdx;
         position = builder.position;
+        donorBoundary = builder.donorBoundary;
+        acceptorBoundary = builder.acceptorBoundary;
     }
 
     public static SplicingLocationData outside() {
-        return newBuilder()
+        return builder()
                 .setSplicingPosition(SplicingPosition.OUTSIDE)
                 .setExonIndex(-1)
                 .setIntronIndex(-1)
                 .build();
     }
 
-    public static Builder newBuilder() {
+    public static Builder builder() {
         return new Builder();
+    }
+
+    public Optional<GenomePosition> getDonorBoundary() {
+        return Optional.ofNullable(donorBoundary);
+    }
+
+    public Optional<GenomePosition> getAcceptorBoundary() {
+        return Optional.ofNullable(acceptorBoundary);
     }
 
     public int getIntronIdx() {
@@ -53,7 +68,7 @@ public class SplicingLocationData {
          */
         EXON,
         /**
-         * Variant is located in intron and  it does not change sequence of splice donor or acceptor sites.
+         * Variant is located in intron and it does not change sequence of splice donor or acceptor sites.
          */
         INTRON,
 
@@ -79,6 +94,8 @@ public class SplicingLocationData {
 
         private SplicingPosition position;
 
+        private GenomePosition donorBoundary, acceptorBoundary;
+
         private Builder() {
         }
 
@@ -94,6 +111,16 @@ public class SplicingLocationData {
 
         public Builder setSplicingPosition(SplicingPosition position) {
             this.position = position;
+            return this;
+        }
+
+        public Builder setDonorBoundary(GenomePosition donorBoundary) {
+            this.donorBoundary = donorBoundary;
+            return this;
+        }
+
+        public Builder setAcceptorBoundary(GenomePosition acceptorBoundary) {
+            this.acceptorBoundary = acceptorBoundary;
             return this;
         }
 
