@@ -4,15 +4,16 @@ import com.google.common.collect.ImmutableMap;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
 /**
- *
+ * POJO for results of splicing analysis.
  */
 public class SplicingPathogenicityData {
 
-    private static final SplicingPathogenicityData EMPTY = SplicingPathogenicityData.newBuilder().build();
+    private static final SplicingPathogenicityData EMPTY = SplicingPathogenicityData.builder().build();
 
-    private final ImmutableMap<ScoringStrategy, Double> scoresMap;
+    private final ImmutableMap<String, Double> scoresMap;
 
     private SplicingPathogenicityData(Builder builder) {
         scoresMap = ImmutableMap.copyOf(builder.scoresMap);
@@ -22,15 +23,15 @@ public class SplicingPathogenicityData {
         return EMPTY;
     }
 
-    public static Builder newBuilder() {
+    public static Builder builder() {
         return new Builder();
     }
 
-    public ImmutableMap<ScoringStrategy, Double> getScoresMap() {
+    public ImmutableMap<String, Double> getScoresMap() {
         return scoresMap;
     }
 
-    public double getOrDefault(ScoringStrategy strategy, double defaultValue) {
+    public double getOrDefault(String strategy, double defaultValue) {
         return scoresMap.getOrDefault(strategy, defaultValue);
     }
 
@@ -41,20 +42,18 @@ public class SplicingPathogenicityData {
                 .orElse(Double.NaN);
     }
 
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof SplicingPathogenicityData)) return false;
-
+        if (o == null || getClass() != o.getClass()) return false;
         SplicingPathogenicityData that = (SplicingPathogenicityData) o;
-
-        return scoresMap != null ? scoresMap.equals(that.scoresMap) : that.scoresMap == null;
-
+        return Objects.equals(scoresMap, that.scoresMap);
     }
 
     @Override
     public int hashCode() {
-        return scoresMap != null ? scoresMap.hashCode() : 0;
+        return Objects.hash(scoresMap);
     }
 
     @Override
@@ -66,18 +65,18 @@ public class SplicingPathogenicityData {
 
     public static final class Builder {
 
-        private Map<ScoringStrategy, Double> scoresMap;
+        private Map<String, Double> scoresMap;
 
         private Builder() {
             this.scoresMap = new HashMap<>();
         }
 
-        public Builder putAllScores(Map<ScoringStrategy, Double> scoresMap) {
+        public Builder putAllScores(Map<String, Double> scoresMap) {
             this.scoresMap.putAll(scoresMap);
             return this;
         }
 
-        public Builder putScore(ScoringStrategy strategy, double score) {
+        public Builder putScore(String strategy, double score) {
             this.scoresMap.put(strategy, score);
             return this;
         }
