@@ -1,23 +1,27 @@
 package org.monarchinitiative.threes.core;
 
-import com.google.common.collect.ImmutableMap;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import de.charite.compbio.jannovar.data.ReferenceDictionary;
 import de.charite.compbio.jannovar.data.ReferenceDictionaryBuilder;
-import org.monarchinitiative.threes.core.calculators.sms.FileSMSParser;
+import org.monarchinitiative.threes.core.calculators.ic.SplicingInformationContentCalculator;
 import org.monarchinitiative.threes.core.calculators.sms.SMSCalculator;
-import org.monarchinitiative.threes.core.calculators.sms.SMSParser;
+import org.monarchinitiative.threes.core.data.ic.InputStreamBasedPositionalWeightMatrixParser;
+import org.monarchinitiative.threes.core.data.ic.SplicingPositionalWeightMatrixParser;
+import org.monarchinitiative.threes.core.data.ic.SplicingPwmData;
+import org.monarchinitiative.threes.core.data.sms.FileSMSParser;
+import org.monarchinitiative.threes.core.data.sms.SMSParser;
 import org.monarchinitiative.threes.core.model.SplicingParameters;
-import org.monarchinitiative.threes.core.reference.GenomeCoordinatesFlipper;
+import org.monarchinitiative.threes.core.reference.allele.AlleleGenerator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import javax.sql.DataSource;
 import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Map;
 
 /**
  *
@@ -30,7 +34,7 @@ public class TestDataSourceConfig {
      */
     @Bean
     public DataSource dataSource() {
-        String jdbcUrl = "jdbc:h2:mem:pbga;INIT=CREATE SCHEMA IF NOT EXISTS SPLICING";
+        String jdbcUrl = "jdbc:h2:mem:threes";
         final HikariConfig config = new HikariConfig();
         config.setUsername("sa");
         config.setDriverClassName("org.h2.Driver");
@@ -40,8 +44,17 @@ public class TestDataSourceConfig {
     }
 
     @Bean
+    public SplicingParameters splicingParameters() {
+        return PojosForTesting.makeSplicingParameters();
+    }
+
+    @Bean
     public ReferenceDictionary referenceDictionary() {
         ReferenceDictionaryBuilder builder = new ReferenceDictionaryBuilder();
+
+        builder.putContigID("chr1", 1);
+        builder.putContigName(1, "chr1");
+        builder.putContigLength(1, 10_000);
 
         builder.putContigID("chr2", 2);
         builder.putContigName(2, "chr2");
@@ -54,26 +67,25 @@ public class TestDataSourceConfig {
     }
 
     @Bean
-    public SplicingParameters splicingParameters() {
-        return SplicingParameters.builder()
-                .setDonorExonic(3)
-                .setDonorIntronic(6)
-                .setAcceptorExonic(2)
-                .setAcceptorIntronic(25)
-                .build();
+    public SplicingInformationContentCalculator splicingInformationContentCalculator(SplicingPwmData splicingPwmData) {
+        return new SplicingInformationContentCalculator(splicingPwmData);
     }
 
     @Bean
-    public Map<String, Integer> contigLengthMap() {
-        return ImmutableMap.<String, Integer>builder()
-                .put("chr1", 10000)
-                .put("chr2", 20000)
-                .build();
+    public SplicingPwmData splicingPwmData(SplicingPositionalWeightMatrixParser splicingPositionalWeightMatrixParser) {
+        return splicingPositionalWeightMatrixParser.getSplicingPwmData();
     }
 
     @Bean
-    public GenomeCoordinatesFlipper genomeCoordinatesFlipper(Map<String, Integer> contigLengthMap) {
-        return new GenomeCoordinatesFlipper(contigLengthMap);
+    public SplicingPositionalWeightMatrixParser splicingPositionalWeightMatrixParser() throws IOException {
+        try (InputStream inputStream = Files.newInputStream(Paths.get(TestDataSourceConfig.class.getResource("data/ic/spliceSites.yaml").getPath()))) {
+            return new InputStreamBasedPositionalWeightMatrixParser(inputStream);
+        }
+    }
+
+    @Bean
+    public AlleleGenerator alleleGenerator(SplicingParameters splicingParameters) {
+        return new AlleleGenerator(splicingParameters);
     }
 
     @Bean
@@ -83,7 +95,7 @@ public class TestDataSourceConfig {
 
     @Bean
     public SMSParser smsParser() throws IOException {
-        Path path = Paths.get(TestDataSourceConfig.class.getResource("calculators/sms/septamer-scores.tsv").getPath());
+        Path path = Paths.get(TestDataSourceConfig.class.getResource("data/sms/septamer-scores.tsv").getPath());
         return new FileSMSParser(path);
     }
 }
