@@ -1,43 +1,34 @@
 package org.monarchinitiative.threes.core.calculators.sms;
 
-import org.junit.jupiter.api.BeforeAll;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Map;
+import org.monarchinitiative.threes.core.TestDataSourceConfig;
+import org.monarchinitiative.threes.core.data.sms.SMSParser;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.closeTo;
 
+
+@SpringBootTest(classes = TestDataSourceConfig.class)
 class SMSCalculatorTest {
 
     private static final double EPSILON = 0.0005;
 
-    /**
-     * Path to file with real scores
-     */
-    private static Path SEPTAMER_TSV_PATH = Paths.get(SMSCalculatorTest.class.getResource("septamer-scores.tsv").getPath());
-
-    private static Map<String, Double> SEPTA_MAP;
+    @Autowired
+    private SMSParser smsParser;
 
     private SMSCalculator calculator;
 
-    @BeforeAll
-    static void setUpBefore() throws IOException {
-        FileSMSParser parser = new FileSMSParser(SEPTAMER_TSV_PATH);
-        SEPTA_MAP = parser.getSeptamerMap();
-
-    }
 
     @BeforeEach
     void setUp() {
-        calculator = new SMSCalculator(SEPTA_MAP);
+        calculator = new SMSCalculator(smsParser.getSeptamerMap());
     }
 
 
