@@ -1,15 +1,17 @@
 package org.monarchinitiative.threes.core.reference.allele;
 
-import org.junit.jupiter.api.BeforeAll;
+import de.charite.compbio.jannovar.data.ReferenceDictionary;
+import de.charite.compbio.jannovar.reference.GenomeInterval;
+import de.charite.compbio.jannovar.reference.GenomePosition;
+import de.charite.compbio.jannovar.reference.GenomeVariant;
+import de.charite.compbio.jannovar.reference.Strand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.monarchinitiative.threes.core.TestDataSourceConfig;
-import org.monarchinitiative.threes.core.model.GenomeCoordinates;
-import org.monarchinitiative.threes.core.model.SequenceInterval;
 import org.monarchinitiative.threes.core.model.SplicingParameters;
-import org.monarchinitiative.threes.core.model.SplicingVariant;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import xyz.ielis.hyperutil.reference.fasta.SequenceInterval;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
@@ -18,35 +20,33 @@ import static org.hamcrest.MatcherAssert.assertThat;
 @SpringBootTest(classes = {TestDataSourceConfig.class})
 class AlleleGeneratorTest {
 
-    private static SequenceInterval donorSi;
+    private SequenceInterval donorSi;
 
-    private static SequenceInterval acceptorSi;
+    private SequenceInterval acceptorSi;
+
+    private GenomePosition anchor;
 
     @Autowired
     private SplicingParameters splicingParameters;
 
-    private AlleleGenerator generator;
+    @Autowired
+    private ReferenceDictionary referenceDictionary;
 
-    @BeforeAll
-    static void setUpBefore() {
-        donorSi = SequenceInterval.of(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(95)
-                        .setEnd(110)
-                        .setStrand(true)
-                        .build(),
-                "TGATGgtaggtgaaa");
-        acceptorSi = SequenceInterval.of(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(70)
-                        .setEnd(110)
-                        .setStrand(true)
-                        .build(),
-                "atggcaaacactgttccttctctctttcagGTGGCCCTGC");
-    }
+    private AlleleGenerator generator;
 
     @BeforeEach
     void setUp() {
+        donorSi = SequenceInterval.builder()
+                .interval(new GenomeInterval(referenceDictionary, Strand.FWD, 1, 93, 110))
+                .sequence("CGTGATGgtaggtgaaa")
+                .build();
+
+        acceptorSi = SequenceInterval.builder()
+                .interval(new GenomeInterval(referenceDictionary, Strand.FWD, 1, 70, 110))
+                .sequence("atggcaaacactgttccttctctctttcagGTGGCCCTGC")
+                .build();
+
+        anchor = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
         generator = new AlleleGenerator(splicingParameters);
     }
 
@@ -54,272 +54,209 @@ class AlleleGeneratorTest {
 
     @Test
     void simpleSnp() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(100)
-                        .setEnd(101)
-                        .setStrand(true)
-                        .build())
-                .setRef("G")
-                .setAlt("T")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
-        assertThat(allele, is("ATGTtaggt"));
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "g", "t");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
+        assertThat(allele, is("ATGttaggt"));
     }
 
     @Test
     void shortDeletion() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(100)
-                        .setEnd(103)
-                        .setStrand(true)
-                        .build())
-                .setRef("gta")
-                .setAlt("g")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "gta", "g");
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
         assertThat(allele, is("ATGgggtga"));
     }
 
     @Test
     void shortInsertion() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(100)
-                        .setEnd(101)
-                        .setStrand(true)
-                        .build())
-                .setRef("g")
-                .setAlt("gcc")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "g", "gcc");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
         assertThat(allele, is("ATGgcctag"));
     }
 
     @Test
     void insertionAcross3PrimeBoundary() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(104)
-                        .setEnd(105)
-                        .setStrand(true)
-                        .build())
-                .setRef("g")
-                .setAlt("gcc")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 104);
+        final GenomeVariant variant = new GenomeVariant(position, "g", "gcc");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
         assertThat(allele, is("ATGgtaggc"));
     }
 
     @Test
     void deletionAcross3PrimeBoundary() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(104)
-                        .setEnd(107)
-                        .setStrand(true)
-                        .build())
-                .setRef("gtg")
-                .setAlt("g")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 104);
+        final GenomeVariant variant = new GenomeVariant(position, "gtg", "g");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
         assertThat(allele, is("ATGgtagga"));
     }
 
     @Test
     void deletionOfWholeSite() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(96)
-                        .setEnd(106)
-                        .setStrand(true)
-                        .build())
-                .setRef("GATGgtaggt")
-                .setAlt("G")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 96);
+        final GenomeVariant variant = new GenomeVariant(position, "GATGgtaggt", "G");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
         assertThat(allele, is(nullValue()));
     }
 
     @Test
     void deletionSpanningFirstBasesOfDonor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(96)
-                        .setEnd(99)
-                        .setStrand(true)
-                        .build())
-                .setRef("GAT")
-                .setAlt("G")
-                .build();
-        // reference is TGATGgtaggtgaaa
-        final String allele = generator.getDonorSiteWithAltAllele(100, variant, donorSi);
-        assertThat(allele, is("TGGgtaggt"));
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 94);
+        final GenomeVariant variant = new GenomeVariant(position, "GTGAT", "G");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
+        assertThat(allele, is("CGGgtaggt"));
     }
 
+    @Test
+    void mismatchInContigsForDonor() {
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 2, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "g", "t");
+
+        // reference is CGTGATGgtaggtgaaa
+        final String allele = generator.getDonorSiteWithAltAllele(anchor, variant, donorSi);
+        assertThat(allele, is(nullValue()));
+    }
+
+    @Test
+    void donorRefAllele() {
+        final String donorSeq = generator.getDonorSiteSnippet(anchor, donorSi);
+        // reference is CGTGATGgtaggtgaaa
+        assertThat(donorSeq, is("ATGgtaggt"));
+    }
+
+    @Test
+    void donorRefAlleleIsNullWhenBadInput() {
+        final String donorSeq = generator.getDonorSiteSnippet(new GenomePosition(referenceDictionary, Strand.FWD, 22, 100), donorSi);
+        // reference is CGTGATGgtaggtgaaa
+        assertThat(donorSeq, is(nullValue()));
+    }
+
+    @Test
+    void makeDonorInterval() {
+        final GenomeInterval donor = generator.makeDonorInterval(anchor);
+        assertThat(donor, is(new GenomeInterval(referenceDictionary, Strand.FWD, 1, 97, 106)));
+    }
 
     // --------------------------      ACCEPTOR ALLELE      -----------------------
 
 
     @Test
     void simpleSnpInAcceptor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(100)
-                        .setEnd(101)
-                        .setStrand(true)
-                        .build())
-                .setRef("G")
-                .setAlt("C")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "G", "C");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is("aaacactgttccttctctctttcagCT"));
     }
 
 
     @Test
     void shortDeletionInAcceptor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(97)
-                        .setEnd(100)
-                        .setStrand(true)
-                        .build())
-                .setRef("cag")
-                .setAlt("c")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 97);
+        final GenomeVariant variant = new GenomeVariant(position, "cag", "c");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is("gcaaacactgttccttctctctttcGT"));
     }
 
 
     @Test
     void shortInsertionInAcceptor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(97)
-                        .setEnd(98)
-                        .setStrand(true)
-                        .build())
-                .setRef("c")
-                .setAlt("ctt")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 97);
+        final GenomeVariant variant = new GenomeVariant(position, "c", "ctt");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is("acactgttccttctctctttcttagGT"));
     }
 
 
     @Test
     void insertionAcross3PrimeBoundaryInAcceptor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(100)
-                        .setEnd(101)
-                        .setStrand(true)
-                        .build())
-                .setRef("G")
-                .setAlt("GCC")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "G", "GCC");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is("acactgttccttctctctttcagGCCT"));
     }
 
 
     @Test
     void deletionAcross3PrimeBoundaryInAcceptor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(100)
-                        .setEnd(104)
-                        .setStrand(true)
-                        .build())
-                .setRef("GTGG")
-                .setAlt("G")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "GTGG", "G");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is("aaacactgttccttctctctttcagGC"));
     }
 
 
     @Test
     void deletionOfTheWholeAcceptorSite() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(73)
-                        .setEnd(102)
-                        .setStrand(true)
-                        .build())
-                .setRef("gcaaacactgttccttctctctttcagGT")
-                .setAlt("g")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 73);
+        final GenomeVariant variant = new GenomeVariant(position, "gcaaacactgttccttctctctttcagGT", "g");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is(nullValue()));
     }
 
 
     @Test
     void deletionSpanningFirstBasesOfAcceptor() {
-        final SplicingVariant variant = SplicingVariant.newBuilder()
-                .setCoordinates(GenomeCoordinates.newBuilder()
-                        .setContig("chr1")
-                        .setBegin(73)
-                        .setEnd(76)
-                        .build())
-                .setRef("gca")
-                .setAlt("g")
-                .build();
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 1, 73);
+        final GenomeVariant variant = new GenomeVariant(position, "gca", "g");
+
         // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
-        final String allele = generator.getAcceptorSiteWithAltAllele(100, variant, acceptorSi);
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
         assertThat(allele, is("gaacactgttccttctctctttcagGT"));
     }
 
+    @Test
+    void mismatchInContigsForAcceptor() {
+        final GenomePosition position = new GenomePosition(referenceDictionary, Strand.FWD, 2, 100);
+        final GenomeVariant variant = new GenomeVariant(position, "G", "C");
 
-    // ---------------   static Region class   ---------------
-
+        // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
+        final String allele = generator.getAcceptorSiteWithAltAllele(anchor, variant, acceptorSi);
+        assertThat(allele, is(nullValue()));
+    }
 
     @Test
-    void differenceTo() {
-        final AlleleGenerator.Region first = generator.makeDonorRegion(100);
+    void acceptorRefAllele() {
+        final String acceptorSeq = generator.getAcceptorSiteSnippet(anchor, acceptorSi);
+        // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
+        assertThat(acceptorSeq, is("aaacactgttccttctctctttcagGT"));
+    }
 
-        int diff = first.differenceTo(110, 120);
-        assertThat(diff, is(-5));
+    @Test
+    void acceptorRefAlleleIsNullWhenBadInput() {
+        final String acceptorSeq = generator.getAcceptorSiteSnippet(new GenomePosition(referenceDictionary, Strand.FWD, 22, 100), acceptorSi);
+        // reference is atggcaaacactgttccttctctctttcagGTGGCCCTGC
+        assertThat(acceptorSeq, is(nullValue()));
+    }
 
-        diff = first.differenceTo(120, 120);
-        assertThat(diff, is(-15));
-
-        diff = first.differenceTo(90, 95);
-        assertThat(diff, is(3));
-
-        diff = first.differenceTo(80, 85);
-        assertThat(diff, is(13));
-
-        diff = first.differenceTo(99, 101);
-        assertThat(diff, is(0));
+    @Test
+    void makeAcceptorInterval() {
+        final GenomeInterval acceptor = generator.makeAcceptorInterval(anchor);
+        assertThat(acceptor, is(new GenomeInterval(referenceDictionary, Strand.FWD, 1, 75, 102)));
     }
 }
