@@ -1,11 +1,15 @@
 package org.monarchinitiative.threes.ingest.transcripts;
 
-import de.charite.compbio.jannovar.data.JannovarData;
+import de.charite.compbio.jannovar.reference.TranscriptModel;
 import org.monarchinitiative.threes.ingest.ProgressLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  *
@@ -18,27 +22,35 @@ public class TranscriptsIngestRunner implements Runnable {
 
     private final TranscriptIngestDao dao;
 
-    private final JannovarData jannovarData;
+    private final Collection<TranscriptModel> transcripts;
 
-    public TranscriptsIngestRunner(SplicingCalculator calculator, TranscriptIngestDao dao, JannovarData jannovarData) {
+    public TranscriptsIngestRunner(SplicingCalculator calculator,
+                                   TranscriptIngestDao dao,
+                                   Collection<TranscriptModel> transcripts) {
         this.calculator = calculator;
         this.dao = dao;
-        this.jannovarData = jannovarData;
+        this.transcripts = transcripts;
     }
 
     @Override
     public void run() {
-        LOGGER.info("Processing {} transcripts", jannovarData.getTmByAccession().values().size());
+        LOGGER.info("Processing {} transcripts", transcripts.size());
 
         ProgressLogger progress = new ProgressLogger();
-        int inserted = jannovarData.getTmByAccession().values().parallelStream()
-                .peek(progress.logTotal("Processed {} transcripts"))
-                .map(calculator::calculate)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .map(dao::insertTranscript)
-                .reduce(Integer::sum)
-                .orElse(0);
+        int inserted = 0;
+        final Map<Integer, List<TranscriptModel>> txByChromosome = transcripts.stream()
+                .collect(Collectors.groupingBy(TranscriptModel::getChr));
+        for (Integer chrom : txByChromosome.keySet()) {
+            LOGGER.info("Processing chromosome `{}`", chrom);
+            inserted += txByChromosome.get(chrom).parallelStream()
+                    .peek(progress.logTotal("Processed {} transcripts"))
+                    .map(calculator::calculate)
+                    .filter(Optional::isPresent)
+                    .map(Optional::get)
+                    .map(dao::insertTranscript)
+                    .reduce(Integer::sum)
+                    .orElse(0);
+        }
         LOGGER.info("Inserted {} transcripts", inserted);
     }
 }
