@@ -1,17 +1,31 @@
 package org.monarchinitiative.threes.ingest;
 
+import com.google.common.collect.ImmutableList;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
+import de.charite.compbio.jannovar.data.JannovarData;
 import de.charite.compbio.jannovar.data.ReferenceDictionary;
 import de.charite.compbio.jannovar.data.ReferenceDictionaryBuilder;
-import org.monarchinitiative.threes.core.reference.GenomeCoordinatesFlipper;
+import de.charite.compbio.jannovar.reference.GenomeInterval;
+import de.charite.compbio.jannovar.reference.Strand;
+import de.charite.compbio.jannovar.reference.TranscriptModel;
+import de.charite.compbio.jannovar.reference.TranscriptModelBuilder;
+import org.monarchinitiative.threes.core.calculators.ic.SplicingInformationContentCalculator;
+import org.monarchinitiative.threes.core.data.ic.InputStreamBasedPositionalWeightMatrixParser;
+import org.monarchinitiative.threes.core.data.ic.SplicingPositionalWeightMatrixParser;
+import org.monarchinitiative.threes.core.data.ic.SplicingPwmData;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.core.JdbcTemplate;
+import xyz.ielis.hyperutil.reference.fasta.GenomeSequenceAccessor;
+import xyz.ielis.hyperutil.reference.fasta.GenomeSequenceAccessorBuilder;
 
 import javax.sql.DataSource;
-import java.util.Map;
-import java.util.stream.Collectors;
+import java.io.InputStream;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  *
@@ -19,6 +33,15 @@ import java.util.stream.Collectors;
 @Configuration
 public class TestDataSourceConfig {
 
+
+    @Bean
+    public SplicingPwmData splicingPwmData() {
+        return SplicingPwmData.builder()
+                .setDonor(PojosForTesting.makeDonorMatrix())
+                .setAcceptor(PojosForTesting.makeAcceptorMatrix())
+                .setParameters(PojosForTesting.makeSplicingParameters())
+                .build();
+    }
 
     @Bean
     public JdbcTemplate jdbcTemplate(DataSource dataSource) {
@@ -30,10 +53,11 @@ public class TestDataSourceConfig {
      */
     @Bean
     public DataSource dataSource() {
-        String jdbcUrl = "jdbc:h2:mem:splicing;INIT=CREATE SCHEMA IF NOT EXISTS SPLICING";
+//        String jdbcUrl = "jdbc:h2:mem:splicing;INIT=CREATE SCHEMA IF NOT EXISTS SPLICING";
+        String jdbcUrl = "jdbc:h2:mem:splicing";
         final HikariConfig config = new HikariConfig();
         config.setUsername("sa");
-//        config.setPassword("");
+        config.setPassword("sa");
         config.setDriverClassName("org.h2.Driver");
         config.setJdbcUrl(jdbcUrl);
 
@@ -55,11 +79,58 @@ public class TestDataSourceConfig {
     }
 
     @Bean
-    public GenomeCoordinatesFlipper genomeCoordinatesFlipper(ReferenceDictionary referenceDictionary) {
-        final Map<String, Integer> contigLengths = referenceDictionary.getContigIDToName().keySet().stream()
-                .collect(Collectors.toMap(idx -> referenceDictionary.getContigIDToName().get(idx),
-                        idx -> referenceDictionary.getContigIDToLength().get(idx)));
-        return new GenomeCoordinatesFlipper(contigLengths);
+    public List<TranscriptModel> transcriptModels(ReferenceDictionary referenceDictionary) {
+        List<TranscriptModel> models = new ArrayList<>();
+
+        // adam
+        final TranscriptModelBuilder adam = new TranscriptModelBuilder();
+        adam.setGeneSymbol("ADAM");
+        adam.setAccession("adam");
+        adam.setStrand(Strand.FWD);
+        adam.setTXRegion(new GenomeInterval(referenceDictionary, Strand.FWD, 2, 10_000, 20_000));
+        adam.setCDSRegion(new GenomeInterval(referenceDictionary, Strand.FWD, 2, 11_000, 19_000));
+        adam.addExonRegion(new GenomeInterval(referenceDictionary, Strand.FWD, 2, 10_000, 12_000));
+        adam.addExonRegion(new GenomeInterval(referenceDictionary, Strand.FWD, 2, 14_000, 16_000));
+        adam.addExonRegion(new GenomeInterval(referenceDictionary, Strand.FWD, 2, 18_000, 20_000));
+        models.add(adam.build());
+
+        return models;
+    }
+
+    @Bean
+    public SplicingInformationContentCalculator splicingInformationContentCalculator(SplicingPositionalWeightMatrixParser splicingPositionalWeightMatrixParser) {
+        return new SplicingInformationContentCalculator(splicingPositionalWeightMatrixParser.getSplicingPwmData());
+    }
+
+    @Bean
+    public SplicingPositionalWeightMatrixParser splicingPositionalWeightMatrixParser() throws Exception {
+        try (InputStream is = TestDataSourceConfig.class.getResourceAsStream("spliceSites.yaml")) {
+            return new InputStreamBasedPositionalWeightMatrixParser(is);
+        }
+    }
+
+    @Bean
+    public JannovarData jannovarData(ReferenceDictionary referenceDictionary, List<TranscriptModel> transcriptModels) {
+        return new JannovarData(referenceDictionary, ImmutableList.copyOf(transcriptModels));
+    }
+
+    /**
+     * This accessor operates on small FASTA file that contains these regions:
+     * - chr2:1-100_000
+     * - chr3:1-200_000
+     * <p>
+     * The names of contigs are `chr2`, and `chr3`, though
+     */
+    @Bean
+    public GenomeSequenceAccessor genomeSequenceAccessor() {
+        Path fasta = Paths.get(TestDataSourceConfig.class.getResource("chr2chr3_small.fa").getPath());
+        Path fastaFai = Paths.get(TestDataSourceConfig.class.getResource("chr2chr3_small.fa.fai").getPath());
+        Path fastaDict = Paths.get(TestDataSourceConfig.class.getResource("chr2chr3_small.fa.dict").getPath());
+        return GenomeSequenceAccessorBuilder.builder()
+                .setFastaPath(fasta)
+                .setFastaFaiPath(fastaFai)
+                .setFastaDictPath(fastaDict)
+                .build();
     }
 
 }
